+++ bricks/bloc/__brick__/{{name.lowerCase()}}/data/data.dart
@@ -1,0 +1,1 @@
+export 'repository/{{name.lowerCase()}}_repository.dart';