+++ bricks/bloc/__brick__/{{name.lowerCase()}}/ui/ui.dart
@@ -1,0 +1,3 @@
+export 'bloc/{{name.lowerCase()}}_bloc.dart';
+export 'view/view.dart';
+export 'widgets/widgets.dart';