+++ bricks/bloc/__brick__/{{name.lowerCase()}}/ui/view/view.dart
@@ -1,0 +1,2 @@
+export '{{name.lowerCase()}}_page.dart';
+export '{{name.lowerCase()}}_view.dart';