+++ bricks/bloc/__brick__/{{name.lowerCase()}}/ui/view/{{name.lowerCase()}}_view.dart
@@ -1,0 +1,10 @@
+part of '{{name.lowerCase()}}_page.dart';
+
+class {{name.pascalCase()}}View extends StatelessWidget {
+  const {{name.pascalCase()}}View({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return {{name.pascalCase()}}Content();
+  }
+}