+++ bricks/bloc/__brick__/{{name.lowerCase()}}/ui/widgets/{{name.lowerCase()}}_content.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class {{name.pascalCase()}}Content extends StatelessWidget {
+  const {{name.pascalCase()}}Content({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const SizedBox.shrink();
+  }
+}