+++ bricks/bloc/__brick__/{{name.lowerCase()}}/{{name.lowerCase()}}.dart
@@ -1,0 +1,2 @@
+export 'data/data.dart';
+export 'ui/ui.dart';