+++ lib/functions/index.dart
@@ -1,2 +1,3 @@
+export 'bloc.dart';
 export 'getx.dart';
 export 'simple.dart';