+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 # 0.3.0
+
 * Updated dependencies and SDK version
 * Added more documentation
 
 # 0.2.0
-* Changed project strcture to use GitPath for bricks
+
+* Changed project structure to use GitPath for bricks
 * Converted the simple folder structure to use bricks
 * Added ability to specifiy a custom path within the lib folder
 
 # 0.1.1
+
 * Added the ability to support GetX as state management solution
 * Changed internal working to use mason bricks
 
@@ -16,3 +19,10 @@
 * Ability to generate screen, widgets
 * Ability to generate data models
 * Ability to generate data repository, services and providers
+
+# 1.0.0
+
+* Added the ability to support BLoC as state management solution
+* Ability to generate bloc, event, state classes
+* Ability to generate data repository, UI directory which includes view, widgets, bloc directories
+* Adding the availability to update pubspec.yaml, adding the necessary packages